pyproject.toml
[project]
name = "khabaru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "python-dotenv>=1.0.1",
    "django-allauth[socialaccount]>=64.1.0",
    "huey>=2.5.1",
    "redis>=5.0.8",
    "pytz>=2024.1",
    "gunicorn>=23.0.0",
    "django-crispy-forms>=2.3",
    "crispy-tailwind>=1.0.3",
    "pillow>=10.4.0",
    "django-custom-error-views>=0.2.5",
    "psycopg[binary,pool]>=3.2.1",
    "django-storages[s3]>=1.14.4",
    "django-cotton>=0.9.32",
    "django-debug-toolbar ==4.2.0",
    "django-browser-reload ==1.12.1",
    "requests>=2.32.3",
    "aider-chat>=0.46.1",
    "django-allauth-ui>=1.4.1",
    "django-cors-headers>=4.4.0",
    "django-taggit>=5.0.1",
]


[tool.uv]
dev-dependencies = [
    "ruff>=0.6.3",
    "aider-chat[browser] ==0.46.1",
]
